re-added conflict management with version 16.04
The package for nova-docker was first set to version 16.04 as YY.MM.
Later on, we decided to follow the OpenStack versioning scheme so the
same sources were versioned as 12.0.0. Since it is a lower version
number we had to use "Breaks", "Replaces" and "Conflicts" statements in
debian/control and the rpm spec file. This way a user could update the
package, even if for the package manager it was not an update since
version went from 16.04 to 12.0.0.

We mistakenly deleted the "Breaks", "Replaces" and "Conflicts"
statements for version 13.0.0. These statements are needed as long we
are in a version < 16.04 since a user with the first version 16.04 can
not update the package with newer versions (12.0.0, 13.0.0) otherwise.

In this commit we re-introduce these statements.